+++ .idea/misc.xml
@@ -5,6 +5,7 @@
       <map>
         <entry key="..\:/Users/con_e/AndroidStudioProjects/LeavingtheNest/app/src/main/res/layout/activity_main.xml" value="0.1" />
         <entry key="..\:/Users/con_e/AndroidStudioProjects/LeavingtheNest/app/src/main/res/layout/content_main.xml" value="0.1" />
+        <entry key="..\:/Users/con_e/AndroidStudioProjects/LeavingtheNest/app/src/main/res/menu/menu_main.xml" value="0.1" />
         <entry key="app/src/main/res/layout/content_main.xml" value="0.24230072463768115" />
         <entry key="app/src/main/res/menu/menu_main.xml" value="0.2453125" />
       </map>
