+++ app/src/main/AndroidManifest.xml
@@ -11,14 +11,18 @@
         android:theme="@style/Theme.LeavingtheNest">
         <activity
             android:name=".MainActivity"
-            android:exported="true"
-            android:label="@string/app_name"
-            android:theme="@style/Theme.LeavingtheNest.NoActionBar">
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
+        </activity>
+        <activity
+            android:name=".FirebaseChatActivity"
+            android:exported="true"
+            android:label="@string/app_name"
+            android:theme="@style/Theme.LeavingtheNest.NoActionBar">
         </activity>
     </application>
