+++ app/src/main/res/layout/content_firebase_chat.xml
@@ -5,8 +5,8 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".MainActivity"
-    tools:showIn="@layout/activity_main">
+    tools:context=".FirebaseChatActivity"
+    tools:showIn="@layout/activity_firebase">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerView"
