+++ app/src/main/res/menu/menu_main.xml
@@ -1,16 +1,30 @@
+<?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     tools:context="com.example.sprint.firebasefun.MainActivity">
     <item
-        android:id="@+id/action_settings"
-        android:orderInCategory="100"
-        android:title="@string/action_settings"
-        app:showAsAction="never" />
-
+        android:id="@+id/menuNavigate"
+        android:icon="@android:drawable/ic_menu_more"
+        android:title="@string/navigate"
+        app:showAsAction="always">
+    <menu>
     <item
-        android:id="@+id/action_signout"
-        android:orderInCategory="100"
-        android:title="Sign Out"
-        app:showAsAction="never" />
+        android:id="@+id/aboutUs"
+        android:icon="@android:drawable/ic_menu_preferences"
+        android:title="@string/about_us"
+        app:showAsAction="ifRoom" />
+    <item
+        android:id="@+id/chatActivity"
+        android:icon="@android:drawable/ic_menu_send"
+        android:title="@string/ask_questions"
+        app:showAsAction="ifRoom" />
+    <item
+        android:id="@+id/watchVideos"
+        android:icon="@android:drawable/presence_video_online"
+        android:title="@string/watch_videos"
+        app:showAsAction="ifRoom"
+        />
+    </menu>
+    </item>
 </menu>
