+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">LeavingtheNest</string>
+    <string name="app_name">Leaving The Nest</string>
     <string name="action_settings">Settings</string>
     <!-- Strings used for fragments for navigation -->
     <string name="first_fragment_label">First Fragment</string>
@@ -9,4 +9,8 @@
 
     <string name="hello_first_fragment">Hello first fragment</string>
     <string name="hello_second_fragment">Hello second fragment. Arg: %1$s</string>
+    <string name="about_us">About us</string>
+    <string name="navigate">Navigate</string>
+    <string name="watch_videos">Watch Videos</string>
+    <string name="ask_questions">Ask Questions</string>
 </resources>
